refactor: changed type for Entity's dimension and Position's coordinates

+++ src/main/scala/model/entities/World.scala
@@ -96,9 +96,7 @@ object World:
   def apply(difficulty: Int = 1, maxLevel: Int = 4): World = WorldImpl(difficulty, maxLevel)
 
   private class WorldImpl(val difficulty: Int, val maxLevel: Int) extends World:
-    
     private val bstIterator = BstGenerator.generate(difficulty, maxLevel)
-
     private val idPlayer = "player"
     private val openDoor = "door_open"
     private var _currentLevel = 1
@@ -149,4 +147,4 @@ object World:
    * @param x coordinate
    * @param y coordinate
    */
-  case class Position(x: Float, y: Float)
+  case class Position(x: Int, y: Int)
